requirements.txt

# Requirements for Satellite Imagery Land Use Classification

# Core
torch==2.1.2
torchvision==0.16.2
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
opencv-python==4.8.1.78

# Experiment Tracking
mlflow==2.9.2

# Serving
fastapi==0.104.1
uvicorn==0.24.0.post1

# Visualization (optional)
streamlit==1.28.2
matplotlib==3.8.2
seaborn==0.13.0

# Data Download
requests==2.31.0
tqdm==4.66.1

# For EuroSAT dataset
Pillow==10.1.0
